feat(core): implement relationship field introspection for widget auto-detection
- Added \`classify_field\` to \`hyperadmin.core.fields\` to introspect Pydantic model fields.
- Supports detection of Enum, list[Enum], SQLAlchemy/SQLModel relationships (FK and M2M), and static list[str].
- Added support for Python 3.10+ union syntax (\`T | None\`) in field classification.
- Added \`preload_fields\` to \`AdminOptions\` for explicit control over choice preloading.
- Exported \`classify_field\` from \`hyperadmin.core\`.
- Added comprehensive unit tests for field classification.

Part of: #159
Depends on: #160
