+++ src/hyperadmin/core/__init__.py
@@ -1,6 +1,6 @@
 """Core components for HyperAdmin."""
 
 from .choices import ChoiceItem, ChoicesProvider, SelectFieldMeta
-from .discovery import classify_field
+from .fields import classify_field
 
 __all__ = ["ChoiceItem", "ChoicesProvider", "SelectFieldMeta", "classify_field"]
